Move NavTreeItemColumns out of the way.
